+++ R/build_scaledic_skeleton.R
@@ -4,11 +4,11 @@
 #' @param nrows Number of empty rows added to data frame.
 #'
 #' @return When 'filename' is not empty, it writes an Excel file with an empty
-#'   template of a dic file. When 'filename' is empty (ie. ""), returns a data
+#'   template of a dic file. When 'filename' is NA, returns a data
 #'   frame.
 #'
 #' @examples
-#' build_scaledic_skeleton("")
+#' build_scaledic_skeleton(NA, nrows = 3)
 #' @export
 
 build_scaledic_skeleton <- function(filename = "dic_template.xlsx", nrows = 0) {
@@ -17,8 +17,7 @@ build_scaledic_skeleton <- function(filename = "dic_template.xlsx", nrows = 0) {
   out <- as.data.frame(out)
   names(out) <- unlist(.dic_file[-1])
 
-  if (filename == "") return(out)
+  if (is.na(filename)) return(out)
   openxlsx::write.xlsx(out, filename)
   cat(filename, " written at ", getwd(), "\n")
 }
-